Rearranged components. Also, trying to set up a navbar so users can choose between signing up and logging.
