+++ frontend/src/components/WelcomePage.jsx
@@ -1,8 +1,16 @@
+import { Container, Grid } from '@mui/material';
+import { useState } from 'react';
+import LoginForm from './forms/LoginForm';
+import SignUpForm from './forms/SignUpForm';
+import WelcomeHeader from './WelcomeHeader';
+
 const WelcomePage = () => {
+  const [signUp, setSignUp] = useState(false);
   return (
-    <div>
-      <div></div>
-    </div>
+    <Container>
+      <WelcomeHeader setSignUp={setSignUp} />
+      {signUp ? <SignUpForm /> : <LoginForm />}
+    </Container>
   );
 };
